Hice una serie de cosas:
1. Avancé con el BE del siguiente caso de uso a implementar ( CUS 14 - Modificar estado de pista de aeropuerto )
2. Avancé con la implementación de acceso a datos del caso de uso mencionado anteriormente.
3. Avancé con la implementación de la lógica de negocio del caso de uso anteior.
4. Realicé las implementaciones correspondientes para llevar a cabo el caso de uso en la capa de presentación. ( frmDisponibilidad )

Por lo tanto, recomiendo a que se empiece a realizar las pruebas debidas para evaluar el caso de uso, en donde se verifique que el estado de una pista realmente cambia correctamente.

+++ trunk/Construcción/UPC.Trabajo.KBHit.DALC/SqlTUUADALC.cs
@@ -131,7 +131,7 @@ namespace UPC.Trabajo.KBHit.DALC
                     obj.CodTUUA = dr.GetInt32(dr.GetOrdinal("CodTUUA"));
                     obj.CodBoleto = dr.GetInt32(dr.GetOrdinal("CodBoleto"));
                     obj.CodVuelo = dr.GetInt32(dr.GetOrdinal("CodVuelo"));
-                    obj.Impuesto = dr.GetFloat(dr.GetOrdinal("Impuesto"));
+                    obj.Impuesto = (float)dr.GetDouble(dr.GetOrdinal("Impuesto"));
                     obj.Fecha = dr.GetDateTime(dr.GetOrdinal("Fecha"));
                     DALCFactory FabricaDALC = DALCFactory.getFabrica(DALCFactory.SQL);
                     TipoVueloDALC objTipoVueloDALC = FabricaDALC.getTipoVuelo();
